refact: eval on bgg circuit
